@@ -25,6 +25,7 @@ const FarmingManage = lazy(() => import('craPages/farming/manage'));
 const TokenSale = lazy(() => import('craPages/sale'));
 const Ssov = lazy(() => import('craPages/ssov'));
 const SsovManage = lazy(() => import('craPages/ssov/Manage'));
+// const SsovPutsManage = lazy(() => import('craPages/ssov/Manage/Puts'));
 const Nfts = lazy(() => import('craPages/nfts'));
 const CommunityNfts = lazy(() => import('craPages/nfts/community'));
 const DiamondPepesNfts = lazy(() => import('craPages/nfts/diamondpepes'));
@@ -32,6 +33,7 @@ const Oracles = lazy(() => import('craPages/oracles'));
 // const Portfolio = lazy(() => import('pages/portfolio'));
 // const Options = lazy(() => import('pages/options'));
 // const Pools = lazy(() => import('pages/pools'));
+// const PoolsMargin = lazy(() => import('craPages/pools/margin'));
 // const PoolsManage = lazy(() => import('pages/pools/manage'));
 // const PoolsVolume = lazy(() => import('pages/pools/volume'));
 // const TestnetFaucet = lazy(() => import('pages/testnet-faucet'));
@@ -53,7 +55,7 @@ const SsovRoutes = () => {
     <SsovProvider>
       <Routes>
         <Route path="*" element={<Ssov />} />
-        <Route path="manage/:asset" element={<SsovManage />} />
+        <Route path=":type/:asset" element={<SsovManage />} />
       </Routes>
     </SsovProvider>
   );
@@ -75,19 +77,16 @@ function AppRoutes() {
   //   return (
   //     <BrowserRouter forceRefresh={false}>
   //       <Suspense fallback={<PageLoader />}>
-  //         <Switch>
-  //           <Route path="/" component={Options} exact />
-  //           <Route path="/pools" component={Pools} exact />
-  //           <Route path="/pools/manage" component={PoolsManage} exact />
-  //           <Route path="/pools/volume" component={PoolsVolume} exact />
-  //           <Route path="/portfolio" component={Portfolio} exact />
-  //           <Route path="/faucet" component={TestnetFaucet} exact />
-  //           <Route path="/swap" component={Swap} exact />
-  //           <SsovProvider>
-  //             <Route path="/ssov" component={Ssov} exact />
-  //             <Route path="/ssov/manage" component={SsovManage} exact />
-  //           </SsovProvider>
-  //         </Switch>
+  //         <Route path="/" element={<Options />} />
+  //         <Route path="/pools" element={<Pools />} />
+  //         <Route path="/pools/manage" element={<PoolsManage />} />
+  //         <Route path="/pools/volume" element={<PoolsVolume />} />
+  //         <Route path="/pools/margin" element={<PoolsMargin />} />
+  //         <Route path="/portfolio" element={<Portfolio />} />
+  //         <Route path="/faucet" element={<TestnetFaucet />} />
+  //         <Route path="/swap" element={<Swap />} />
+  //         <Route path="ssov/*" element={<SsovRoutes />} />
+  //         <Route path="*" element={<Error statusCode={404} />} />
   //       </Suspense>
   //     </BrowserRouter>
   //   );
