@@ -1,4 +1,4 @@
-const ArrowRightIcon = ({ className }) => (
+const ArrowRightIcon = ({ className, fill = '#3E3E3E' }) => (
   <svg
     className={className}
     width="10"
@@ -8,7 +8,7 @@ const ArrowRightIcon = ({ className }) => (
   >
     <path
       d="M.917 5.583h6.516L4.586 8.43a.588.588 0 0 0 0 .828.58.58 0 0 0 .822 0l3.845-3.844a.58.58 0 0 0 0-.822L5.414.742a.58.58 0 1 0-.822.822l2.84 2.853H.918A.585.585 0 0 0 .333 5c0 .32.263.583.584.583Z"
-      fill="#3E3E3E"
+      fill={fill}
     />
   </svg>
 );
