@@ -11,6 +11,8 @@ export interface Data {
   maxLpDeposits: BigNumber;
   mintPrice: BigNumber;
   totalDeposits: BigNumber;
+  lpSupply: BigNumber;
+  lpReserves: [BigNumber, BigNumber, number];
 }
 
 export interface DiamondPepeContextInterface {
@@ -22,11 +24,13 @@ export interface DiamondPepeContextInterface {
 
 export const initialData: DiamondPepeContextInterface = {
   data: {
-    isDepositPeriod: false,
+    isDepositPeriod: true,
     isFarmingPeriod: false,
     maxLpDeposits: BigNumber.from('0'),
     mintPrice: BigNumber.from('0'),
     totalDeposits: BigNumber.from('0'),
+    lpReserves: [BigNumber.from('0'), BigNumber.from('0'), 0],
+    lpSupply: BigNumber.from('0'),
   },
   userData: { deposits: BigNumber.from('0'), minted: false },
 };
