@@ -10,12 +10,19 @@ import EpochSelector from 'craPages/pools/components/EpochSelector';
 import PoolCard from './components/PoolCard';
 import PoolSelectorButton from './components/PoolSelectorButton';
 import VolumeCard from './components/VolumeCard';
+import LendingCard from './components/LendingCard';
 
 import { PoolsContext, PoolsProvider } from 'contexts/Pools';
 
 import getNextFriday from 'utils/date/getNextFriday';
 
 import styles from './styles.module.scss';
+
+enum SelectedPool {
+  POOLS,
+  VOLUME_POOLS,
+  LENDING_POOLS,
+}
 
 function Pools() {
   const {
@@ -24,7 +31,9 @@ function Pools() {
     selectedEpoch,
     epochInitTime,
   } = useContext(PoolsContext);
-  const [volumePools, setVolumePools] = useState<boolean>(false);
+  const [selectedPool, setSelectedPool] = useState<SelectedPool>(
+    SelectedPool.POOLS
+  );
 
   useEffect(() => {
     window.scrollTo(0, 0);
@@ -53,7 +62,7 @@ function Pools() {
         <Box className="flex mt-5">
           <PoolSelectorButton
             title={'Pools'}
-            isSelected={!volumePools}
+            isSelected={selectedPool === SelectedPool.POOLS}
             startIcon={
               <Box className="h-10 w-10 rounded-full flex items-center bg-stieglitz">
                 <Typography variant="h5" className="ml-4 mb-1">
@@ -66,14 +75,21 @@ function Pools() {
                 ? `WETH, WBTC + ${baseAssetsOptionPoolSdks.length - 2} More`
                 : `WETH, WBTC`
             }
-            onClick={() => setVolumePools(false)}
+            onClick={() => setSelectedPool(SelectedPool.POOLS)}
           />
           <PoolSelectorButton
             startIcon={<img src={'/assets/usdt.svg'} alt="USDT" />}
             title={'Volume Pools'}
             subtitle={'USDT'}
-            isSelected={volumePools}
-            onClick={() => setVolumePools(true)}
+            isSelected={selectedPool === SelectedPool.VOLUME_POOLS}
+            onClick={() => setSelectedPool(SelectedPool.VOLUME_POOLS)}
+          />
+          <PoolSelectorButton
+            startIcon={<img src={'/assets/usdt.svg'} alt="USDT" />}
+            title={'Lending Pools'}
+            subtitle={'USDT'}
+            isSelected={selectedPool === SelectedPool.LENDING_POOLS}
+            onClick={() => setSelectedPool(SelectedPool.LENDING_POOLS)}
           />
         </Box>
         {/* <TimePeriodSelector className="mt-1.5" /> */}
@@ -100,8 +116,10 @@ function Pools() {
         </Box>
         <Box className="mt-5 flex-1 flex overflow-auto">
           <Box className="flex min-h-0 space-x-6">
-            {volumePools ? (
+            {selectedPool === SelectedPool.VOLUME_POOLS ? (
               <VolumeCard />
+            ) : selectedPool === SelectedPool.LENDING_POOLS ? (
+              <LendingCard />
             ) : (
               baseAssetsOptionPoolSdks.map((baseAssetsOptionPoolSdk, index) => (
                 <PoolCard
