@@ -1,5 +1,5 @@
-import { useCallback, useContext, useState, useEffect, useMemo } from 'react';
-import { BigNumber, ethers } from 'ethers';
+import { useCallback, useContext, useState, useMemo } from 'react';
+import { BigNumber } from 'ethers';
 import TableRow from '@material-ui/core/TableRow';
 import TableCell from '@material-ui/core/TableCell';
 import Box from '@material-ui/core/Box';
@@ -9,7 +9,7 @@ import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 
 import { WalletContext } from 'contexts/Wallet';
-import { SsovProperties, SsovData } from 'contexts/Ssov';
+import { SsovContext } from 'contexts/Ssov';
 
 import CustomButton from 'components/UI/CustomButton';
 import Typography from 'components/UI/Typography';
@@ -31,8 +31,6 @@ interface ExerciseTableDataProps {
   totalPremiumsEarned: BigNumber;
   isSettleable: boolean;
   isPastEpoch: boolean;
-  ssovProperties: SsovProperties;
-  ssovData: SsovData;
 }
 
 const DIALOGS = {
@@ -51,24 +49,22 @@ const ExerciseTableData = (props: ExerciseTableDataProps) => {
     pnlAmount,
     isSettleable,
     isPastEpoch,
-    ssovProperties,
-    ssovData,
   } = props;
 
   const { contractAddresses, signer } = useContext(WalletContext);
+  const { ssovData, ssovEpochData, selectedEpoch } = useContext(SsovContext);
 
   const tokenSymbol =
-    SSOV_MAP[ssovProperties.tokenName].tokenSymbol === 'BNB'
+    SSOV_MAP[ssovData.tokenName].tokenSymbol === 'BNB'
       ? 'vBNB'
-      : SSOV_MAP[ssovProperties.tokenName].tokenSymbol;
+      : SSOV_MAP[ssovData.tokenName].tokenSymbol;
 
-  const { selectedEpoch } = ssovProperties;
-  const { epochStrikes, isEpochExpired } = ssovData;
+  const { epochStrikes, isEpochExpired } = ssovEpochData;
 
   const [dialogState, setDialogState] = useState({
     open: false,
     type: 'SETTLE',
-    ssovProperties: ssovProperties,
+    ssovData,
   });
 
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
@@ -83,18 +79,18 @@ const ExerciseTableData = (props: ExerciseTableDataProps) => {
       setDialogState({
         open: true,
         type: 'TRANSFER',
-        ssovProperties: ssovProperties,
+        ssovData,
       }),
-    [ssovProperties]
+    [ssovData]
   );
   const handleSettle = useCallback(
     () =>
       setDialogState({
         open: true,
         type: 'SETTLE',
-        ssovProperties: ssovProperties,
+        ssovData,
       }),
-    [ssovProperties]
+    [ssovData]
   );
 
   const handleClickMenu = useCallback(
@@ -132,7 +128,7 @@ const ExerciseTableData = (props: ExerciseTableDataProps) => {
         open={dialogState.open}
         handleClose={handleClose}
         strikeIndex={strikeIndex}
-        ssovProperties={ssovProperties}
+        ssovData={ssovData}
         token={tokenSymbol}
         settleableAmount={settleableAmount}
         className="rounded-xl"
@@ -141,7 +137,7 @@ const ExerciseTableData = (props: ExerciseTableDataProps) => {
         <Box className="h-12 flex flex-row items-center">
           <Box className="flex flex-row h-8 w-8 mr-2">
             <img
-              src={SSOV_MAP[ssovProperties.tokenName].imageSrc}
+              src={SSOV_MAP[ssovData.tokenName].imageSrc}
               alt={tokenSymbol}
             />
           </Box>
