@@ -1,11 +1,13 @@
 import { useState, useEffect, useCallback, useContext, useMemo } from 'react';
 import {
   OptionPool,
+  ERC20,
   OptionPoolBroker__factory,
   ERC20__factory,
   VolumePool__factory,
   Dopex__factory,
   Router__factory,
+  Margin__factory,
 } from '@dopex-io/sdk';
 import { useFormik } from 'formik';
 import noop from 'lodash/noop';
@@ -21,6 +23,7 @@ import isZeroAddress from 'utils/contracts/isZeroAddress';
 import { getWeeklyPool, getMonthlyPool } from 'utils/contracts/getPool';
 import oneEBigNumber from 'utils/math/oneEBigNumber';
 import getTimePeriod from 'utils/contracts/getTimePeriod';
+import getOptionPoolId from 'utils/contracts/getOptionPoolId';
 
 import useSendTx from 'hooks/useSendTx';
 
@@ -58,16 +61,29 @@ const useOptionPurchase = () => {
   const [userVolumePoolFunds, setUserVolumePoolFunds] = useState('0');
   const [txError, setTxError] = useState('');
 
-  const { selectedOptionData, expiry } = useContext(OptionsContext);
+  const { selectedOptionData, expiry, marginData } = useContext(OptionsContext);
   const {
     selectedBaseAsset,
     selectedBaseAssetContract,
     selectedBaseAssetDecimals,
     usdtContract,
     userAssetBalances,
+    usdtDecimals,
   } = useContext(AssetsContext);
   const { provider, accountAddress, contractAddresses, signer } =
     useContext(WalletContext);
+
+  const { collaterals } = marginData;
+  const marginAvailable = collaterals.length > 0;
+  const maxLeverage = useMemo(
+    () =>
+      marginData.maxLeverage !== '0' && marginData.minLeverage !== '0'
+        ? BigNumber.from(marginData.maxLeverage)
+            .div(marginData.minLeverage)
+            .toString()
+        : '0',
+    [marginData.maxLeverage, marginData.minLeverage]
+  );
 
   const sendTx = useSendTx();
 
@@ -135,6 +151,10 @@ const useOptionPurchase = () => {
       amount: 1,
       useVolumePoolFunds: false,
       delegate: false,
+      margin: false,
+      collateralAmount: 1,
+      collateralIndex: 0,
+      leverage: 1,
     },
     validate: (values) => {
       const errors: any = {};
@@ -154,9 +174,8 @@ const useOptionPurchase = () => {
         ).lte(getContractReadableAmount(values.amount, 18))
       ) {
         errors.amount = 'Not enough liquidity';
-        // This else if can be deleted when the new router (that allows volume pool funds to be used) is deployed
-      } else if (values.useVolumePoolFunds && values.delegate) {
-        errors.amount = 'Please enable auto exercise after purchasing';
+      } else if (values.margin && values.delegate) {
+        errors.amount = 'Auto exercise cannot be enabled with margin';
       } else if (
         values.useVolumePoolFunds &&
         BigNumber.from(userVolumePoolFunds).lt(
@@ -171,6 +190,22 @@ const useOptionPurchase = () => {
         )
       ) {
         errors.amount = 'Not enough balance in wallet';
+      } else if (values.margin && !marginAvailable) {
+        errors.amount = 'Margin trading not available';
+      } else if (
+        values.margin &&
+        values.collateralIndex >= collaterals.length
+      ) {
+        errors.amount = 'Invalid collateral selected';
+      } else if (values.margin && values.collateralAmount < 0) {
+        errors.amount = 'Invalid collateral amount';
+      } else if (
+        values.margin &&
+        (values.leverage >
+          Number(marginData.maxLeverage) / Number(marginData.minLeverage) ||
+          values.leverage < 1)
+      ) {
+        errors.amount = 'Invalid leverage';
       }
 
       return errors;
@@ -215,6 +250,38 @@ const useOptionPurchase = () => {
     poolState.weeklyPool,
   ]);
 
+  const collateralValue = useMemo(
+    () =>
+      formik.values.collateralIndex < collaterals.length
+        ? BigNumber.from(
+            collaterals[formik.values.collateralIndex]?.price ?? '0'
+          )
+            .mul(
+              getContractReadableAmount(
+                formik.values.collateralAmount,
+                collaterals[formik.values.collateralIndex]?.decimals ?? 0
+              )
+            )
+            .div(ethers.utils.parseUnits('1', usdtDecimals))
+            .toString()
+        : '0',
+    [formik, collaterals, usdtDecimals]
+  );
+
+  const requiredCollateralValue = useMemo(
+    () =>
+      formik.values.collateralIndex < collaterals.length
+        ? BigNumber.from(totalPrice)
+            .mul(
+              collaterals[formik.values.collateralIndex]
+                ?.collateralizationRatio ?? '0'
+            )
+            .div('100')
+            .toString()
+        : totalPrice,
+    [totalPrice, collaterals, formik]
+  );
+
   const handleChange = useCallback(
     async (e) => {
       formik.setFieldValue('amount', Number(e.target.value));
@@ -228,14 +295,47 @@ const useOptionPurchase = () => {
     [formik]
   );
 
+  const handleCollateralAmount = useCallback(
+    async (e) => {
+      formik.setFieldValue('collateralAmount', Number(e.target.value));
+    },
+    [formik]
+  );
+
+  const handleLeverage = useCallback(
+    async (e) => {
+      formik.setFieldValue('leverage', Number(e.target.value));
+    },
+    [formik]
+  );
+
+  const handleCollateralIndex = useCallback(
+    (e) => formik.setFieldValue('collateralIndex', Number(e)),
+    [formik]
+  );
+
   const purchaseOptions = useCallback(async () => {
     if (!usdtContract || !signer) return;
     setTxError('');
     try {
-      const amount = await getContractReadableAmount(
+      const amount = getContractReadableAmount(
         formik.values.amount,
         selectedBaseAssetDecimals
       );
+
+      const collateralAmount = formik.values.margin
+        ? getContractReadableAmount(
+            formik.values.collateralAmount,
+            collaterals[formik.values.collateralIndex].decimals
+          )
+        : BigNumber.from('0');
+      const leverage = formik.values.margin
+        ? BigNumber.from(
+            Math.round(
+              Number(formik.values.leverage) * Number(marginData.minLeverage)
+            ).toString()
+          )
+        : BigNumber.from('0');
 
       const strike = BigNumber.from(selectedOptionData.strikePrice).mul(
         STRIKE_PRECISION
@@ -247,6 +347,23 @@ const useOptionPurchase = () => {
 
       const router = Router__factory.connect(contractAddresses.Router, signer);
 
+      const margin = contractAddresses.Margin
+        ? Margin__factory.connect(contractAddresses.Margin, signer)
+        : undefined;
+
+      const approveToken: ERC20 = formik.values.margin
+        ? ERC20__factory.connect(
+            collaterals[formik.values.collateralIndex].token,
+            signer
+          )
+        : ERC20__factory.connect(usdtContract.address, signer);
+      const approveContractAddress = formik.values.margin
+        ? margin.address
+        : router.address;
+      const approveAmount = formik.values.margin ? collateralAmount : amount;
+
+      let timePeriod: 'weekly' | 'monthly' | undefined;
+
       if (
         weeklyPool &&
         (await weeklyPool.contract.isValidExpiry(expiry / 1000)) &&
@@ -254,35 +371,7 @@ const useOptionPurchase = () => {
           await weeklyPool.contract.getCurrentEpoch()
         ))
       ) {
-        const allowance = await usdtContract.allowance(
-          accountAddress,
-          router.address
-        );
-
-        if (!amount.lt(allowance)) {
-          const signerUsdtContract = ERC20__factory.connect(
-            usdtContract.address,
-            signer
-          );
-          await sendTx(signerUsdtContract.approve(router.address, MAX_VALUE));
-        }
-
-        const usdtBalance = await usdtContract.balanceOf(accountAddress);
-
-        await sendTx(
-          router.purchaseOption(usdtBalance, {
-            useVolumePoolFunds: formik.values.useVolumePoolFunds,
-            isPut: isPut,
-            delegate: formik.values.delegate,
-            strike: strike,
-            expiry: expiry / 1000,
-            amount: amount,
-            timePeriod: getTimePeriod('weekly'),
-            baseAssetAddress: selectedBaseAssetContract.address,
-            quoteAssetAddress: usdtContract.address,
-            to: accountAddress,
-          })
-        );
+        timePeriod = 'weekly';
       } else if (
         monthlyPool &&
         (await monthlyPool.contract.isValidExpiry(expiry / 1000)) &&
@@ -290,39 +379,55 @@ const useOptionPurchase = () => {
           await monthlyPool.contract.getCurrentEpoch()
         ))
       ) {
-        const allowance = await usdtContract.allowance(
-          accountAddress,
-          monthlyPool.address
-        );
-
-        if (!amount.lt(allowance)) {
-          const signerUsdtContract = ERC20__factory.connect(
-            usdtContract.address,
-            signer
-          );
-          await sendTx(
-            signerUsdtContract.approve(monthlyPool.address, MAX_VALUE)
-          );
-        }
-
-        const usdtBalance = await usdtContract.balanceOf(accountAddress);
-
-        await sendTx(
-          router.purchaseOption(usdtBalance, {
-            useVolumePoolFunds: formik.values.useVolumePoolFunds,
-            isPut: isPut,
-            delegate: formik.values.delegate,
-            strike: strike,
-            expiry: expiry / 1000,
-            amount: amount,
-            timePeriod: getTimePeriod('monthly'),
-            baseAssetAddress: selectedBaseAssetContract.address,
-            quoteAssetAddress: usdtContract.address,
-            to: accountAddress,
-          })
-        );
+        timePeriod = 'monthly';
       } else {
         setTxError('Not enough liquidity');
+      }
+
+      if (timePeriod) {
+        const allowance = await approveToken.allowance(
+          accountAddress,
+          approveContractAddress
+        );
+
+        if (!approveAmount.lt(allowance)) {
+          await sendTx(approveToken.approve(approveContractAddress, MAX_VALUE));
+        }
+
+        if (formik.values.margin) {
+          await margin.openMarginPosition({
+            useVolumePoolFunds: formik.values.useVolumePoolFunds,
+            isPut: isPut,
+            amount: amount,
+            leverage: leverage,
+            strike: strike,
+            expiry: expiry / 1000,
+            collateralAmount: collateralAmount,
+            collateral: collaterals[formik.values.collateralIndex].token,
+            optionPoolId: getOptionPoolId(
+              selectedBaseAssetContract.address,
+              usdtContract.address,
+              timePeriod
+            ),
+          });
+        } else {
+          const usdtBalance = await usdtContract.balanceOf(accountAddress);
+
+          await sendTx(
+            router.purchaseOption(usdtBalance, {
+              useVolumePoolFunds: formik.values.useVolumePoolFunds,
+              isPut: isPut,
+              delegate: formik.values.delegate,
+              strike: strike,
+              expiry: expiry / 1000,
+              amount: amount,
+              timePeriod: getTimePeriod(timePeriod),
+              baseAssetAddress: selectedBaseAssetContract.address,
+              quoteAssetAddress: usdtContract.address,
+              to: accountAddress,
+            })
+          );
+        }
       }
     } catch (err) {
       setTxError(`Something went wrong. Error: ${parseError(err)}`);
@@ -341,6 +446,8 @@ const useOptionPurchase = () => {
     poolState,
     contractAddresses,
     signer,
+    collaterals,
+    marginData,
     sendTx,
   ]);
 
@@ -357,6 +464,11 @@ const useOptionPurchase = () => {
     poolState.weeklyBasePoolAssets,
     selectedOptionData.strikePrice,
   ]);
+
+  const handleMargin = useCallback(
+    (e) => formik.setFieldValue('margin', e.target.checked),
+    [formik]
+  );
 
   const handleUseVolumePool = useCallback(
     (e) => formik.setFieldValue('useVolumePoolFunds', e.target.checked),
@@ -404,9 +516,18 @@ const useOptionPurchase = () => {
     selectedOptionData,
     totalPrice,
     userAssetBalances,
+    handleMargin,
     handleUseVolumePool,
     handleDelegate,
+    handleCollateralIndex,
+    handleCollateralAmount,
+    handleLeverage,
     userVolumePoolFunds,
+    maxLeverage,
+    collaterals,
+    marginAvailable,
+    collateralValue,
+    requiredCollateralValue,
   };
 };
 
